# api/openApi.yaml
openapi: 3.0.3
info:
  title: Issues API
  version: 1.0.0
paths:
  /issues:
    get:
      summary: List all issues
      responses:
        "200":
          description: List of issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IssueDto"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      summary: Create a new issue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIssueDto"
      responses:
        "201":
          description: Issue created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /issues/{id}:
    get:
      summary: Get issue by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Issue found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: Delete issue by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Issue delete
          content:
            application/json:
              schema:
                type: object
                properties:
                  messaage:
                    type: string
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /issues/status:
    get:
      summary: Get all issue statuses
      responses:
        "200":
          description: Issue status
          content:
            application/json:
              schema:
                type: array
                items:
                $ref: "#/components/schemas/IssueStatusDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /issues/{id}/status:
    get:
      summary: Get issue status by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Issue status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueStatusDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

components:
  responses:
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    IssueDto:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        owner:
          type: string
        repo:
          type: string
        number:
          type: number
    CreateIssueDto:
      type: object
      properties:
        url:
          type: string
    IssueStatusDto:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
        url:
          type: string
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string